backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.1.2
slowapi==0.1.9

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Utilities
python-dotenv==1.0.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.3
gunicorn==21.2.0